backend/requirements.txt
asgiref==3.7.2
asttokens==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
decorator==5.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.11
django-allauth==0.61.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.14.0
exceptiongroup==1.3.0
executing==2.2.0
idna==3.10
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.1.7
nose==1.3.7
oauthlib==3.2.2
parso==0.8.4
pathlib==1.0.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
